tegra: pcie: use integer for root_bus_nr
The drivers internal root_bus_nr used to be u8 which lead to a wrong
error detection in bus_to_port. Bus number can be -1 in case bus is
not scanned yet. Thanks to James pointing that out.

+++ arch/arm/mach-tegra/pcie.c
@@ -289,7 +289,7 @@
 
 struct tegra_pcie_port {
 	int			index;
-	u8			root_bus_nr;
+	int			root_bus_nr;
 	void __iomem		*base;
 
 	bool			link_up;
